chore: Moves the stylesheets folder to the assets folder
