+++ superset-frontend/src/SqlLab/main.less
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-@import '../../stylesheets/less/variables.less';
+@import '../assets/stylesheets/less/variables.less';
 
 body {
   min-height: ~'max(100vh, 500px)'; // Set a min height so the gutter is always visible when resizing
