+++ superset-frontend/src/assets/stylesheets/less/cosmo/bootswatch.less
