+++ superset-frontend/src/assets/stylesheets/less/cosmo/variables.less
