+++ superset-frontend/src/assets/stylesheets/less/variables.less
