+++ superset-frontend/src/assets/stylesheets/reactable-pagination.less
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-@import '../stylesheets/less/variables.less';
+@import './less/variables.less';
 
 .reactable-pagination td {
   padding: 15px 0 0 0 !important;
