+++ superset-frontend/src/assets/stylesheets/superset.less
@@ -579,6 +579,6 @@ hr {
   border-top: 1px solid @gray-light;
 }
 .ace_gutter-cell.ace_error {
-  background-image: url('../src/assets/images/icons/error_solid_small_red.svg') !important;
+  background-image: url('../images/icons/error_solid_small_red.svg') !important;
   background-position: -2px center !important;
 }
