+++ superset-frontend/src/components/FilterableTable/FilterableTableStyles.less
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-@import '../../../stylesheets/less/variables.less';
+@import '../../assets/stylesheets/less/variables.less';
 
 .ReactVirtualized__Grid__innerScrollContainer {
   border: 1px solid @gray-light;
