+++ superset-frontend/src/dashboard/stylesheets/builder-sidepane.less
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-@import '../../../stylesheets/less/variables.less';
+@import '../../assets/stylesheets/less/variables.less';
 
 .dashboard-builder-sidepane {
   flex: 0 0 @builder-pane-width;
