+++ superset-frontend/src/dashboard/stylesheets/index.less
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-@import './../../../stylesheets/less/variables.less';
+@import '../../assets/stylesheets/less/variables.less';
 
 @import './builder.less';
 @import './builder-sidepane.less';
