+++ superset-frontend/src/explore/main.less
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-@import '../../stylesheets/less/variables.less';
+@import '../assets/stylesheets/less/variables.less';
 
 .scrollbar-container {
   position: relative;
