+++ superset-frontend/src/theme.ts
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import '../stylesheets/superset.less';
+import './assets/stylesheets/superset.less';
 
 // Importing Antd under its own stylesheet to prevent unintentional theming.
-import '../stylesheets/antd/index.less';
+import './assets/stylesheets/antd/index.less';
