+++ superset-frontend/src/visualizations/FilterBox/FilterBox.less
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-@import '../../../stylesheets/less/variables.less';
+@import '../../assets/stylesheets/less/variables.less';
 
 .dashboard .filter_box .slice_container > div:not(.alert) {
   padding-top: 0;
